Major refactor, split into Viewer
Having a combined viewer and search tool created several issues,
most around document permissions. When DocuTube tried to load on
a View Only document, it would fail completely because the search
tools tried to initialize the Cache on the document.

This commit strips out the search tool. Search functionality will be
maintained in a separate repo and published under a separate Addon
to cut down on errors caused by permissions.